Fixed an issue with proxy_pass and proxy_pass_match parameters
Creating a virtual host with the current layout causes an error in Apache 2.4.6, as Apache thinks that the parameter is a command. For example, when doing:

```
proxy_pass_match => [
    {
      'path' => '^/(.*\.php(/.*)?)$',
      'url' => 'fcgi://127.0.0.1:9000/var/www/html/phoenix/web/$1',
      'params' => {'timeout'=>60}
    }
```

Apache outputs an error saying that the `timeout=60` command is not recognized. The reason is that when this template is rendered, the output looks something like this:

```
ProxyPassMatch ^/(.*\.php(/.*)?)$ fcgi://127.0.0.1:9000/var/www/html/phoenix/web/$1
 timeout=60
```

As you might have noticed, there is a line break, and then a space, before the `timeout=60 parameter.

This is what I'm trying to fix in this pull request.